pyproject.toml
[project]
name = "embedkit"
version = "0.1.0"
description = "A simple toolkit for generating vector embeddings across multiple providers and models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.7.0",
    "cohere>=5.15.0",
    "colpali-engine>=0.3.0,<0.4.0",
    "pillow>=11.2.1",
    "torch<=2.5",
    "transformers",
]
authors = [
    {name = "JP Hwang", email = "me@jphwang.com"},
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/embedkit"]
